PlatformParser
