+++ com.sap.cloud.lm.sl.mta/src/main/java/com/sap/cloud/lm/sl/mta/parsers/v2/PlatformParser.java
@@ -2,27 +2,41 @@ package com.sap.cloud.lm.sl.mta.parsers.v2;
 
 import static com.sap.cloud.lm.sl.mta.handlers.v2.Schemas.PLATFORM;
 
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 
 import com.sap.cloud.lm.sl.common.ParsingException;
-import com.sap.cloud.lm.sl.common.util.ListUtil;
 import com.sap.cloud.lm.sl.mta.model.v2.Platform;
 import com.sap.cloud.lm.sl.mta.model.v2.ModuleType;
 import com.sap.cloud.lm.sl.mta.model.v2.ResourceType;
 import com.sap.cloud.lm.sl.mta.model.v2.Platform.Builder;
+import com.sap.cloud.lm.sl.mta.parsers.ListParser;
+import com.sap.cloud.lm.sl.mta.parsers.ModelParser;
 import com.sap.cloud.lm.sl.mta.schema.MapElement;
 
-public class PlatformParser extends com.sap.cloud.lm.sl.mta.parsers.v1.PlatformParser {
+public class PlatformParser extends ModelParser<Platform> {
 
+    protected static final String PROCESSED_OBJECT_NAME = "MTA platform type";
+
+    public static final String NAME = "name";
+    public static final String DESCRIPTION = "description";
+    public static final String VERSION = "version";
+    public static final String MODULE_TYPES = "module-types";
+    public static final String PROPERTIES = "properties";
+    public static final String RESOURCE_TYPES = "resource-types";
     public static final String PARAMETERS = "parameters";
 
+    protected final Set<String> usedModuleTypeNames = new HashSet<>();
+    protected final Set<String> usedResourceTypeNames = new HashSet<>();
+
     public PlatformParser(Map<String, Object> source) {
-        super(PLATFORM, source);
+        this(PLATFORM, source);
     }
 
     protected PlatformParser(MapElement schema, Map<String, Object> source) {
-        super(schema, source);
+        super(PROCESSED_OBJECT_NAME, schema, source);
     }
 
     @Override
@@ -36,20 +50,46 @@ public class PlatformParser extends com.sap.cloud.lm.sl.mta.parsers.v1.PlatformP
         return builder.build();
     }
 
-    protected List<ModuleType> getModuleTypes2() {
-        return ListUtil.cast(getModuleTypes1());
+    protected String getName() {
+        return getStringElement(NAME);
     }
 
-    @Override
+    protected String getVersion() {
+        return getStringElement(VERSION);
+    }
+
+    protected String getDescription() {
+        return getStringElement(DESCRIPTION);
+    }
+
+    protected Map<String, Object> getProperties() {
+        return getMapElement(PROPERTIES);
+    }
+
+    protected List<ModuleType> getModuleTypes2() {
+        return getListElement(MODULE_TYPES, new ListParser<ModuleType>() {
+            @Override
+            protected ModuleType parseItem(Map<String, Object> map) {
+                return getModuleTypeParser(map).setUsedValues(usedModuleTypeNames)
+                    .parse();
+            }
+        });
+    }
+
     protected ModuleTypeParser getModuleTypeParser(Map<String, Object> source) {
         return new ModuleTypeParser(source);
     }
 
     protected List<ResourceType> getResourceTypes2() {
-        return ListUtil.cast(getResourceTypes1());
+        return getListElement(RESOURCE_TYPES, new ListParser<ResourceType>() {
+            @Override
+            protected ResourceType parseItem(Map<String, Object> map) {
+                return getResourceTypeParser(map).setUsedValues(usedResourceTypeNames)
+                    .parse();
+            }
+        });
     }
 
-    @Override
     protected ResourceTypeParser getResourceTypeParser(Map<String, Object> source) {
         return new ResourceTypeParser(source);
     }
